Adiciona entidade ranking e a implementa nas camada de dados e negócio.

+++ Modulo 19/WhoWantsToBeAMillionaire/WhoWantsToBeAMillionaire.Core/Models/AwardModel.cs
@@ -6,11 +6,11 @@ namespace WhoWantsToBeAMillionaire.Core.Models;
 public class AwardModel
 {
     public int QuestionNumber { get; private set; }
-    public decimal Correct { get; private set; }
-    public decimal Stop { get; private set; }
-    public decimal Wrong { get; private set; }
+    public int Correct { get; private set; }
+    public int Stop { get; private set; }
+    public int Wrong { get; private set; }
 
-    public AwardModel(int questioNumber, decimal correct, decimal stop, decimal wrong)
+    public AwardModel(int questioNumber, int correct, int stop, int wrong)
     {
         QuestionNumber = questioNumber;
         Correct = correct;
